% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance4vec.R
\name{cov_function}
\alias{cov_function}
\title{cov_function}
\usage{
cov_function(x, y, method = c("pearson", "spearman"))
}
\arguments{
\item{x}{input vector}

\item{y}{input vector, with the same dimension of x}

\item{method}{a character string that indicates specific form of correlation, 
either pearson, kendall or spearman}
}
\value{
spearman covariance given x and y
}
\description{
Gets pearson or spearman covariance given two vectors
}
\examples{
cov_function(c(1,2,3,4,5),c(2,1,0,3,4),method="spearman") 
}
