% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_glucose_simple.R
\docType{data}
\name{data_glucose_simple}
\alias{data_glucose_simple}
\title{Simple glucose dataset}
\format{
A data frame with 6 rows and 3 variables:
\describe{
  \item{subject}{index number}
  \item{age}{age of subject, in years}
  \item{glucose.level}{glucose level of subject, in standard unit}
  ...
}
}
\source{
\url{https://www.statisticshowto.com/probability-and-statistics/regression-analysis/find-a-linear-regression-equation/}
}
\usage{
data_glucose_simple
}
\description{
A very simple dataset containing age and glucose level for 6 people
}
\keyword{datasets}
