% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{fit_linear_regression}
\alias{fit_linear_regression}
\title{Fit simple linear regression model}
\usage{
fit_linear_regression(x_name, y_name, data)
}
\arguments{
\item{x_name}{independent variable name}

\item{y_name}{dependent variable name}

\item{data}{data.frame that contains both variables}
}
\description{
This function takes three parameters and fits simple linear regression model.
The function asks for independent variable name (string), dependent variable name (string), and a data frame
(data.frame) that contains both independent and dependent variables. If either variable name is not found in
the data frame, this function will return an error message. Otherwise, it will compute slope and intercept,
calculate standard error, R-squared, and adjusted R-squared values, and perform t-test and f-test on the
independent variable. This function will return a list containing all results.
}
\note{
Please note that this function should not be called by user. It is an internal function used for fitting. User should call linear_regression() instead.
}
\examples{
fit_linear_regression("age", "glucose.level", data_glucose_simple)

}
