% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happiness_data.R
\docType{data}
\name{happiness_data}
\alias{happiness_data}
\title{Sample data - relationship between income and happiness}
\format{
A data frame with 498 rows and 2 variables:
\describe{
  \item{income}{income}
  \item{happiness}{scale of happiness}
  ...
}
}
\source{
\url{https://www.scribbr.com/statistics/simple-linear-regression/}
}
\usage{
happiness_data
}
\description{
A more complex sample data for simple linear regression package.
}
\keyword{datasets}
