% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_regression}
\alias{linear_regression}
\title{Calculate simple linear regression}
\usage{
linear_regression(x_name, y_name, data)
}
\arguments{
\item{x_name}{independent variable name}

\item{y_name}{dependent variable name}

\item{data}{data.frame that contains both variables}
}
\description{
This function takes three parameters and fits simple linear regression model.
The function asks for independent variable name (string), dependent variable name (string), and a data frame
(data.frame) that contains both independent and dependent variables. If either variable name is not found in
the data frame, this function will return an error message. Otherwise, it will compute slope and intercept,
calculate standard error, R-squared, and adjusted R-squared values, and perform t-test and f-test on the
independent variable. This function will return a lm_result object containing all results. This function and
a series of subsequent functions rely on S3 method to print user-friendly results.
}
\examples{
linear_regression("income", "happiness", happiness_data)
linear_regression("age", "glucose.level", data_glucose_simple)
results <- linear_regression("age", "glucose.level", data_glucose_simple)

}
