% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm2.R
\name{lm2}
\alias{lm2}
\title{lm2}
\usage{
lm2(formula, data, na.action = "omit")
}
\arguments{
\item{formula}{An object of class "formula" that models regression to be fit}

\item{data}{Data frame to perform the linear regression on}

\item{na.action}{Character input that determines how the regression model should handle missing data. Options include 'omit' (remove rows with NA values), 'fail'(halt regression), or 'impute' (replace NA values with column mean). Defaults to omit if excluded from call.}
}
\value{
a list that contains the following values: residuals, rank, fitted.values, df.residual, coefficients, call, model, data.frame, cf, y
}
\description{
Used fit linear regression models
}
\examples{
lm2(formula = fixed.acidity ~ volatile.acidity + citric.acid * residual.sugar, data = wine)
example <- lm2(formula = quality ~ free.sulfur.dioxide:pH + pH, data = wine)
example$residuals
example$coefficients

}
