% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_logit.R
\name{my_logit}
\alias{my_logit}
\title{Logistic regression using IRLS}
\usage{
my_logit(formula, data, max_iter = 25, tol = 1e-06)
}
\arguments{
\item{formula}{Model formula, e.g., y ~ x1 + x2}

\item{data}{A data.frame containing the variables}

\item{max_iter}{Maximum number of iterations}

\item{tol}{Tolerance for convergence}
}
\value{
An object of class "my_logit".
}
\description{
Logistic regression using IRLS
}
\examples{
fit <- my_logit(am ~ mpg + wt, data = mtcars)
summary(fit)
residuals(fit)[1:5]

}
