% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov4vec.R
\name{pearson_cor}
\alias{pearson_cor}
\title{cov4vec}
\usage{
pearson_cor(x, y, method, stat_test)
}
\arguments{
\item{x}{input value}

\item{y}{input value}

\item{method}{input string}

\item{stat_test}{input}
}
\value{
the correlation of x and y
}
\description{
Gets different type of correlation, p-value, test statistics given two vectors
}
\examples{
pearson_cor(c(1,2),c(2,1),method="pearson",stat_test=TRUE)

}
