% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance4vec.R
\name{pearson_cov}
\alias{pearson_cov}
\title{pearson covariance}
\usage{
pearson_cov(x, y, method)
}
\arguments{
\item{x}{input vector}

\item{y}{input vector, with the same dimension of x}

\item{method}{a character string that indicates specific form of correlation, 
either pearson, kendall or spearman}
}
\value{
pearson covariance given x and y
}
\description{
Gets different type of correlation, p-value, test statistics given two vectors
}
\examples{
pearson_cov(c(1,2,3,4,5),c(2,1,0,3,4),method="pearson") 
}
