% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{plot_linear_regression}
\alias{plot_linear_regression}
\title{Plot linear regression line}
\usage{
plot_linear_regression(result, data)
}
\arguments{
\item{result}{result from linear_regression}

\item{data}{original data frame used in linear_regression}
}
\description{
This function uses result acquired by calling linear_regression and original data
to plot both original data points and linear regression line on the same plot.
Original data points will be labeled as black dots, whereas regression line will
be red.
}
\examples{
plot_linear_regression(results, data)

}
