% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm2.R, R/print.lm2.R
\name{print.lm2}
\alias{print.lm2}
\title{lm2}
\usage{
\method{print}{lm2}(m)

\method{print}{lm2}(m)
}
\arguments{
\item{m}{output of lm2 function}

\item{formula}{an object of class "formula" that models regression to be fit}

\item{data}{data frame to perform the linear regression on}

\item{na.action}{character input that determines how the regression model should handle missing data. Options include 'omit' (remove rows with NA values), 'fail'(halt regression), or 'mean impute' (replace NA values with column mean)}
}
\value{
a list that contains the following values: residuals, rank, fitted.values, df.residual, coefficients, call, model, data.frame, cf, y
}
\description{
Used fit linear regression models

Formats printing style for lm2 function
}
