% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{print.lm_result}
\alias{print.lm_result}
\title{Print lm_result object}
\usage{
print(x)

print.lm_result(x)
}
\arguments{
\item{x}{lm_result object}

\item{...}{ignored}
}
\description{
This function uses S3 method to print user-friendly result for linear regression.
This function will return command used to call linear_regression, intercept, and slope
for the simple linear regression.
}
