% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{print.summary.lm_result}
\alias{print.summary.lm_result}
\title{Print summary for lm_result object}
\usage{
summary(x)
}
\arguments{
\item{x}{summary.lm_result object}

\item{...}{ignored}
}
\description{
This function uses S3 method to print summary result of lm_result object. This functions returns
user-friends result contaning test statistics for the linear regression model, including standard
error, t-statistics, P-value, R-squared, adjusted R-squared, F-statistics, and corresponding P-value.
}
