% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summarylm2.R
\name{print.summarylm2}
\alias{print.summarylm2}
\title{print.summarylm2}
\usage{
\method{print}{summarylm2}(m)
}
\arguments{
\item{m}{output of summarylm2 function}
}
\description{
Formats printing style for summary.lm2 function
}
