% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summarylm2.R
\name{print_summarylm2}
\alias{print_summarylm2}
\title{print_summarylm2}
\usage{
print_summarylm2(m)
}
\arguments{
\item{m}{A list output generated by summarylm2 function}
}
\description{
Used to print summary lm2 outputs
}
