% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{summary.lm_result}
\alias{summary.lm_result}
\title{Generate correct summary for lm_result object}
\usage{
summary(x)

summary.lm_result(x)
}
\arguments{
\item{...}{ignored}

\item{x}{lm_result object}
}
\description{
This function uses S3 method to generate correct summary for lm_object.
}
