% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarylm2.R
\name{summarylm2}
\alias{summarylm2}
\title{summarylm2}
\usage{
summarylm2(output)
}
\arguments{
\item{output}{return of of lm2.}
}
\value{
a list that contains the following values: call, residuals, coefficients, sigma, df, r.squared, adj.r.squared, fstatistic, cov.unscaled, F.pval
}
\description{
Summarize fit for linear regression models
}
