% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wine.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine Quality Data}
\format{
An object of class \code{data.frame} with 50 rows and 13 columns.
}
\source{
\href{https://www.kaggle.com/uciml/red-wine-quality-cortez-et-al-2009}{Kaggle}
}
\usage{
data(wine)
}
\description{
Data from red variants of of the Portuguese "Vinho Verde" wine. 50 rows and 12 variables. Contains fixed acidity, volatile acidity,
citric acid, residual sugar, chlorides, free sulfur dioxide, total sulfur dioxide, density, pH,
sulphates, alcohol and quality (score between 0 and 10) variables.
}
\examples{
data(wine)
ph <- wine$pH
}
\keyword{datasets}
