% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov4vec.R
\name{pearson_cor}
\alias{pearson_cor}
\title{cov4vec}
\usage{
pearson_cor(x, y, method, stat_test)
}
\arguments{
\item{x}{input vector}

\item{y}{input vector, with the same dimension of x}

\item{method}{a character string that indicates specific form of correlation, 
either pearson, kendall or spearman}

\item{stat_test}{input of TRUE or FASLE, to indicate the program whether
to perform a test of significance and report p-value}
}
\value{
a list of combination of correlation information and value given x and y
}
\description{
Gets different type of correlation, p-value, test statistics given two vectors
}
\examples{
pearson_cor(c(1,2,3,4,5),c(2,1,0,3,4),method="pearson",stat_test=FALSE) 

}
